+++ packages/twoslash/src/annotations/static.ts
